Provide valid values in contour examples (#37824)
The current contour example values of (0,0) (0,0) do not work. Per the documentation linked in this section, the first number of each pair should be followed by percent sign (%), and the second of each pair should be followed by "st" for relative change or "Hz" for specified pitch, as well as preceded by + or -. This change updates the documentation to provide valid example values.