Include Z-Wave.Me UZB1 in the note about Aeotec compatibility issues
The Z-Wave.Me UZB1 is based on the same HW, as lsusb shows:

0658:0200 Sigma Designs, Inc. Aeotec Z-Stick Gen5 (ZW090) - UZB

It's been reported numerous time to be broken and newer Linux kernel
versions can't work with it at all because of changes in the USB stack.
As this has never been a Pi4 issue, adjust the note accordingly and
refer the Z-Wave.Me stick to it too.

The title of the note was left unchanged to keep links referring to it
still working.